[WIP] Implement FIR filter and noise gate
I'm attempting to get the mic volume to a louder base level, and there's a noise at around 2.5kHz-3kHz about as "loud" as the voice signal. Spectral filtering seems to be the correct answer here, but I figured these might be useful.

The code uses ARM's CMSIS quite a bit, so it's not yet portable. I will work on fixing this after the noise can be filtered appropriately.