Removed all the W25Qx_wakeup()/W25Qx_sleep() pairs from code accessing NVM
Avoid putting the W25Qxxx NVM into sleep after each read to not interfere
with the backup system. From now on the memory remains active during all
the time the radio is powered on. This is not optimal in terms of energy
consumption and has to be fixed in a proper way.