Added User Functions thread
This thread handles a set of user functions. 
Those user functions can be either periodic (with a frequency of 1Hz, 10Hz, 20Hz, 50Hz or 100Hz) or aperiodic (asynchronous). 
Periodic functions will be called periodically depending on their scheduling. Async functions need to be triggered usign user_functions_trigger(id).

The maximum amount of user functions that can be handled have to be defined in hwconfig.h using the macro CONFIG_USER_FUNCTIONS. If this macro is not defined, the user functions thread will not be created.