Audio: stm32_dac: fixed missing restore of output idle level after stream end
On CKS32F407 the external trigger source for a DAC channel needs to be disabled
in order for the writes to the DAC data registers to be effective. This is
coherent with the STM32F4 reference manual, but the STM32 devices update the DAC
output on a register write wether the external trigger is enabled or not.