audio_codec: decodeFunc: added threshold on sample energy
Changed decodeFunc() to decode codec2 samples only if the energy exceeds
a given threshold: this guarantees that no noise is decoded and played
back at the speaker when there is no speech ongoing.
This modification makes the decoded speech much clearer, as there is no
noise in between words or at idle (silence). The idle noise effect is
especially emphasized when 12-bit ADCs are used at the transmitter side.