Update audio_codec.c
Make sure that the `decodeFunc()` only decodes Codec2 data if the encoded frame energy exceeds a certain threshold. This guarantees that no noise is decoded and played back at the speaker. After applying this modification, the decoded speech is much clearer, as there is no noise in between words or at idle (silence). The idle noise effect is especially emphasized when 12-bit ADCs are used at the transmitter side.