chore(ci+devtools): create builds using devcontainer, update devcontainer to reliably execute all build and wrap commands on amd64 env
This change makes it so that the devcontainer in this project is able to build and wrap all supported platforms. There is a pre-existing issue on arm64 devices preventing the openrtx_gd77_wrap and openrtx_dm1801_wrap targets from working, which is being tracked in https://github.com/OpenRTX/OpenRTX/issues/344.
This change also makes it so that CI/CD builds against all supported platforms, utilizing the devcontainer as its environment rather than setting up one separately. This makes it so that the CICD builds and dev local builds are more similar and reduces the maintenance overhead.

Co-authored-by: Peter Buchegger <peter@pbuchegger.at>