m17 meta: send and receive text metadata, send gnss metadata for m17 mode
This commit contains the original work on supporting sending and receiving metadata, both text and gnss, for the M17 mode. This was included as part of a larger PR https://github.com/OpenRTX/OpenRTX/pull/314 which the lead requested be split up, with the first PR focusing on this feature alone.

This change was created by taking https://github.com/turnrye/OpenRTX/commit/23d8034c907a25e1f096eb977448ca35c8e0ac3a and picking over only files, and in some case sections of files, in order for the metadata features to function. It has generally been tested to ensure that the picking was done properly, but further verification of the intended functionality has not taken place as part of this commit.

Co-authored-by: Rick <crs026@yahoo.com>
Co-authored-by: Jim Ancona <jim@anconafamily.com>
Co-authored-by: Wojciech Kaczmarski <w.kaczmarski@teletra.pl>