Core: M17: add receiving of GNSS meta text
This implements the M17 protocol's GNSS text feature. When the user is receives M17 traffic with meta text, it is displayed on the UI.
When the text is too long to show, it scrolls. When there is text received and link data, both are combined on to the same row.

Co-authored-by: Ryan Turner <ryan@turnrye.com>
Co-authored-by: Wojciech Kaczmarski <w.kaczmarski@teletra.pl>
Co-authored-by: Jim Ancona <jim@anconafamily.com>