License: adopt tool for automated header license checking
Skywalking Eyes is a tool for checking and fixing license headers in code. This change adds a configuration for Skywalking Eyes for our project and adds a github action to run it in check mode.

A separate commit is necessary to revise the copyright headers in order for the github action to pass.