Automatic sa8x8-fw flashing and patching
Now the 1.4.1 release of sa8x8-fw is downloaded at every build of the
ttwr21 target, it is converted from s37 to bin to c header file and is
included in the OpenRTX firmware image.
OpenRTX, at runtime detects the band of the radio and patches the
firmware with the appropriate identification string.

TG-553