audio: oversample mic input to improve clarity
Oversample the microphone input. This samples at 8x the base rate, and a
decimation filter then averages the samples back to the normal rate (8
kHz). This feature is not enabled for Module17 or Linux because the
audio is already presumed to be clean there.

Note that this feature only applies to modes that use the MCU for its
audio path, so this only applies to M17 (codec2) currently.

The DMA buffer is heap-allocated at 320 * oversampling samples to
cover the time codec2_encode() blocks the thread. At 8x this uses
~5 KiB of heap.

Co-authored-by: Ryan Turner <ryan@turnrye.com>
Signed-off-by: Ryan Turner <ryan@turnrye.com>